fixes and refactor
