+++ libs/microsoft/__init__.py
@@ -1,31 +1,17 @@
-from libs.microsoft.clients import MSAPIClient, MSAuthClient
-from libs.microsoft.credentials import MSAdminCredentials
+from libs.microsoft.auth import get_ms_auth_client
+from libs.microsoft.client import MSWebAPIClient
+from libs.microsoft.credentials import get_credentials
 
-credentials = None
 client = None
-auth_client = None
-
-
-def get_ms_auth_client():
-    global auth_client
-
-    if auth_client is None:
-        auth_client = MSAuthClient()
-    return auth_client
 
 
 def get_ms_client():
-    global client, credentials
+    global client
 
     if client is None:
-        credentials = MSAdminCredentials.initialize()
-        client = MSAPIClient(token=credentials.access_token)
+        ms_credentials = get_credentials()
+        client = MSWebAPIClient(
+            access_token=ms_credentials.ms_access_token,
+            refresh_token=ms_credentials.ms_refresh_token
+        )
     return client
-
-
-def dispose_microsoft_connections() -> None:
-    global credentials, client, auth_client
-
-    credentials = None
-    client = None
-    auth_client = None
