+++ libs/microsoft/auth/exceptions.py
